DUI - Bring in initial Pixel Navbar animation resources [1/?]
****

This commit just brings in the history and resources from the
fantastic work Jacob did pwning this Google animation here...

https://github.com/USA-
RedDragon/frameworks_base-1/commit/55563047385874559c7a6c5ab25f45d9ab75296c

****

Pixel Navbar, reverse engineered from smali.

Notes:
- This is forcing the Pixel Home button on people,
and it looks ugly with the other nav buttons being hollow

- There MAY still be a few fixes yet to be found in this code
because it WAS a big file to RE from smali. I have yet to find
any, but time is the best tester.

- Shoutout to @bigrushdog for providing a decompiled form used
to verify the smali and as a guideline for the more obsure
smali sections. It was a big help.

- Tagging myself here to see adoption rate. @USA-RedDragon

Change-Id: Ief1ee04b540c553848698362298719e489fb3585